requirements.txt
aiogram>=3.4,<4.0
openai>=1.12,<2.0
python-dotenv>=1.0,<2.0
aiofiles>=23.2,<25.0
pydantic>=2.4,<3.0
